Refactoring apim plugin packages
