--- components/esb-tools/plugins/org.wso2.developerstudio.eclipse.registry.apim.perspective/META-INF/MANIFEST.MF
@@ -1,42 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Perspective
-Bundle-SymbolicName: org.wso2.developerstudio.eclipse.registry.apim.perspective;singleton:=true
-Bundle-Version: 2.6.0.qualifier
-Bundle-Activator: org.wso2.developerstudio.eclipse.registry.apim.perspective.Activator
-Require-Bundle: org.eclipse.ui,
- org.eclipse.core.runtime,
- org.wso2.developerstudio.eclipse.kernel.libraries;bundle-version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.libraries;bundle-version="[4.1.0,7.0.0)"
-Bundle-ActivationPolicy: lazy
-Import-Package: org.eclipse.e4.core.contexts,
- org.eclipse.e4.core.services.events,
- org.eclipse.e4.core.services.log,
- org.eclipse.e4.ui.di,
- org.osgi.service.event,
- org.wso2.developerstudio.appcloud.utils.authentication;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.logging.core;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.platform.core;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.platform.core.event;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.platform.core.templates;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.platform.core.utils;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.platform.ui;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.platform.ui.utils;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.base.core;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.base.interfaces;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.base.logger;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.base.managers;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.base.model;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.base.persistent;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.base.ui.controls;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.base.ui.util;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.core;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.core.exception;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.manager.local.utils;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.manager.remote;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.manager.remote.dialog;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.registry.manager.remote.views;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.utils;version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.utils.file;version="[4.1.0,7.0.0)"
-Export-Package: org.wso2.developerstudio.eclipse.registry.apim.action,
- org.wso2.developerstudio.eclipse.registry.apim.preferance
