+++ components/esb-tools/plugins/org.wso2.integrationstudio.registry.apim.perspective/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.wso2.developerstudio.eclipse.registry.apim.perspective</name>
+	<name>org.wso2.integrationstudio.registry.apim.perspective</name>
 	<comment></comment>
 	<projects>
 	</projects>
