+++ components/esb-tools/plugins/org.wso2.integrationstudio.registry.apim.perspective/META-INF/MANIFEST.MF
@@ -1,0 +1,42 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Perspective
+Bundle-SymbolicName: org.wso2.integrationstudio.registry.apim.perspective;singleton:=true
+Bundle-Version: 8.0.0.qualifier
+Bundle-Activator: org.wso2.integrationstudio.registry.apim.perspective.Activator
+Require-Bundle: org.eclipse.ui,
+ org.eclipse.core.runtime,
+ org.wso2.integrationstudio.kernel.libraries,
+ org.wso2.integrationstudio.libraries
+Bundle-ActivationPolicy: lazy
+Import-Package: org.eclipse.e4.core.contexts,
+ org.eclipse.e4.core.services.events,
+ org.eclipse.e4.core.services.log,
+ org.eclipse.e4.ui.di,
+ org.osgi.service.event,
+ org.wso2.integrationstudio.appcloud.utils.authentication,
+ org.wso2.integrationstudio.logging.core,
+ org.wso2.integrationstudio.platform.core,
+ org.wso2.integrationstudio.platform.core.event,
+ org.wso2.integrationstudio.platform.core.templates,
+ org.wso2.integrationstudio.platform.core.utils,
+ org.wso2.integrationstudio.platform.ui,
+ org.wso2.integrationstudio.platform.ui.utils,
+ org.wso2.integrationstudio.registry.base.core,
+ org.wso2.integrationstudio.registry.base.interfaces,
+ org.wso2.integrationstudio.registry.base.logger,
+ org.wso2.integrationstudio.registry.base.managers,
+ org.wso2.integrationstudio.registry.base.model,
+ org.wso2.integrationstudio.registry.base.persistent,
+ org.wso2.integrationstudio.registry.base.ui.controls,
+ org.wso2.integrationstudio.registry.base.ui.util,
+ org.wso2.integrationstudio.registry.core,
+ org.wso2.integrationstudio.registry.core.exception,
+ org.wso2.integrationstudio.registry.manager.local.utils,
+ org.wso2.integrationstudio.registry.manager.remote,
+ org.wso2.integrationstudio.registry.manager.remote.dialog,
+ org.wso2.integrationstudio.registry.manager.remote.views,
+ org.wso2.integrationstudio.utils,
+ org.wso2.integrationstudio.utils.file
+Export-Package: org.wso2.integrationstudio.registry.apim.action,
+ org.wso2.integrationstudio.registry.apim.preferance
