+++ components/esb-tools/plugins/org.wso2.integrationstudio.registry.apim.perspective/icons/carbon-studio-small-logo.gif
