+++ components/esb-tools/plugins/org.wso2.integrationstudio.registry.apim.perspective/schema/org.wso2.developerstudio.aboutDialog.contributor.exsd
